LayoutTests/media/track/audio-track-configuration.html

<!DOCTYPE html>
<html>
<head>
	<meta charset="utf-8">
	<title>video-track-configuration</title>
	<script src="../video-test.js"></script>
	<script>
		window.addEventListener('load', async event => {
			findMediaElement();
			run('video.src = "../content/audio-tracks.mp4"');
			await waitFor(video, 'canplay');
			await testExpectedEventually('video.audioTracks.length', 0, '>');
			await testExpectedEventually('video.audioTracks[0].configuration', null, '!=');
			await testExpectedEventually('video.audioTracks[0].configuration.codec', 'mp4a.40.2');
			await testExpectedEventually('video.audioTracks[0].configuration.sampleRate', 44100);
			await testExpectedEventually('video.audioTracks[0].configuration.numberOfChannels', 2);
			await testExpectedEventually('video.audioTracks[0].configuration.bitrate', 100000, '>=');
			endTest();
		});
	</script>
</head>
<body>
	<video controls muted></video>
</body>
</html>
